ForLoopsLowering: Move handlers to sub-package and separate files.
