+++ compiler/ir/backend.common/src/org/jetbrains/kotlin/backend/common/lower/loops/HeaderInfo.kt
@@ -9,6 +9,21 @@ package org.jetbrains.kotlin.backend.common.lower.loops
 
 import org.jetbrains.kotlin.backend.common.CommonBackendContext
 import org.jetbrains.kotlin.backend.common.ir.Symbols
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.ArrayIndicesHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.ArrayIterationHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.CharSequenceIndicesHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.CharSequenceIterationHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.CollectionIndicesHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.DefaultIterableHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.DefaultProgressionHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.DownToHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.IndexedGetIterationHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.RangeToHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.ReversedHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.StepHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.StringIterationHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.UntilHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.handlers.WithIndexHandler
 import org.jetbrains.kotlin.backend.common.lower.matchers.IrCallMatcher
 import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.declarations.IrVariable
@@ -188,7 +203,7 @@ internal class IndexedGetHeaderInfo(
  */
 internal class WithIndexHeaderInfo(val nestedInfo: HeaderInfo) : HeaderInfo() {
     // We cannot easily reverse `withIndex()` so we do not attempt to handle it. We would have to start from the last value of the index,
-    // easily calculable (or even impossible) in most cases.
+    // which is not easily calculable (or even impossible) in most cases.
     override fun asReversed(): HeaderInfo? = null
 }
 
@@ -295,7 +310,13 @@ internal class DefaultHeaderInfoBuilder(context: CommonBackendContext, scopeOwne
     HeaderInfoBuilder(context, scopeOwnerSymbol) {
     override val callHandlers = listOf(
         ReversedHandler(context, this),
-        WithIndexHandler(context, NestedHeaderInfoBuilderForWithIndex(context, scopeOwnerSymbol))
+        WithIndexHandler(
+            context,
+            NestedHeaderInfoBuilderForWithIndex(
+                context,
+                scopeOwnerSymbol
+            )
+        )
     )
 
     // NOTE: StringIterationHandler MUST come before CharSequenceIterationHandler.
