+++ compiler/ir/backend.common/src/org/jetbrains/kotlin/backend/common/lower/loops/handlers/RangeToHandler.kt
@@ -1,0 +1,42 @@
+/*
+ * Copyright 2010-2020 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.backend.common.lower.loops.handlers
+
+import org.jetbrains.kotlin.backend.common.CommonBackendContext
+import org.jetbrains.kotlin.backend.common.lower.createIrBuilder
+import org.jetbrains.kotlin.backend.common.lower.loops.ProgressionDirection
+import org.jetbrains.kotlin.backend.common.lower.loops.ProgressionHandler
+import org.jetbrains.kotlin.backend.common.lower.loops.ProgressionHeaderInfo
+import org.jetbrains.kotlin.backend.common.lower.loops.ProgressionType
+import org.jetbrains.kotlin.backend.common.lower.matchers.SimpleCalleeMatcher
+import org.jetbrains.kotlin.ir.builders.irInt
+import org.jetbrains.kotlin.ir.expressions.IrCall
+import org.jetbrains.kotlin.ir.symbols.IrSymbol
+import org.jetbrains.kotlin.ir.types.IrType
+import org.jetbrains.kotlin.util.OperatorNameConventions
+
+/** Builds a [HeaderInfo] for progressions built using the `rangeTo` function. */
+internal class RangeToHandler(private val context: CommonBackendContext, private val progressionElementTypes: Collection<IrType>) :
+    ProgressionHandler {
+
+    override val matcher = SimpleCalleeMatcher {
+        dispatchReceiver { it != null && it.type in progressionElementTypes }
+        fqName { it.pathSegments().last() == OperatorNameConventions.RANGE_TO }
+        parameterCount { it == 1 }
+        parameter(0) { it.type in progressionElementTypes }
+    }
+
+    override fun build(expression: IrCall, data: ProgressionType, scopeOwner: IrSymbol) =
+        with(context.createIrBuilder(scopeOwner, expression.startOffset, expression.endOffset)) {
+            ProgressionHeaderInfo(
+                data,
+                first = expression.dispatchReceiver!!,
+                last = expression.getValueArgument(0)!!,
+                step = irInt(1),
+                direction = ProgressionDirection.INCREASING
+            )
+        }
+}
